feat: move to distroless image, remove cars example
