+++ src/config/csp.ts
