+++ src/config/site.ts
@@ -1,8 +1,8 @@
 export const siteConfig = {
   creator: 'Yukasama',
   description: 'A Next.js Template for your next project.',
-  keywords: ['nextjs-template'],
+  keywords: ['Next.js', 'Next.js-Template'],
   links: { github: 'https://github.com/Yukasama' },
   name: 'Next.js Template',
-  url: 'https://nextjs-template.com',
+  url: 'https://nextjs-template.candlezone.eu',
 };
